+++ src/main/scala/atmos/retries/RetryDSL.scala
@@ -80,9 +80,11 @@ import org.slf4j.{ Logger => Slf4jLogger }
  * Note that the `5.minutes` parameter is an instance of `scala.concurrent.duration.FiniteDuration` and that any
  * instance of this class may be used as a policy in `retryFor`.
  *
- * Finally, a retry policy that never terminates (unless directed to by an error classifier) can be created with
- * `retryForever`:
+ * Finally, a retry policy that immediately terminates can be created with `neverRetry` and a retry policy that never
+ * terminates (unless directed to by an error classifier) can be created with `retryForever`:
  * {{{
+ * implicit val retryPolicy = neverRetry
+ * 
  * implicit val retryPolicy = retryForever
  * }}}
  *
@@ -146,6 +148,9 @@ object RetryDSL {
   // Retry policy factories and extensions.
   //
 
+  /** Creates a new retry policy that immediately terminates. */
+  def neverRetry: RetryPolicy = RetryPolicy(TerminationPolicy.ImmediatelyTerminate)
+
   /** Creates a new default retry policy. */
   def retrying: RetryPolicy = RetryPolicy()
 
@@ -156,7 +161,7 @@ object RetryDSL {
    */
   def retryFor(termination: TerminationPolicy): RetryPolicy = RetryPolicy(termination)
 
-  /** Creates a new default retry policy. */
+  /** Creates a new retry policy that never terminates. */
   def retryForever: RetryPolicy = RetryPolicy(TerminationPolicy.NeverTerminate)
 
   /**